Agent/backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
requests>=2.32.0,<3.0.0
pytest>=8.2.0,<9.0.0
httpx>=0.27.0,<1.0.0
SQLAlchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
slowapi>=0.1.9,<1.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
argon2-cffi>=23.1.0,<24.0.0
pyjwt>=2.8.0,<3.0.0
python-jose>=3.3.0,<4.0.0
python-multipart>=0.0.9,<1.0.0
celery[redis]>=5.4.0,<6.0.0
redis>=5.0.0,<6.0.0
